anolis: sched/fair: fix the sysctl of sched_expel_update_interval
ANBZ: #4814

There is a bug of sched_expel_update_interval sysctl. See the
following:

	# cat /proc/sys/kernel/sched_expel_update_interval
	10    0

There are two values. It's wrong.

The variable of sysctl_sched_expel_update_interval is defined as
unsigned long, and there will be two values when read this sysctl,
because the size of long will be splited into two size of int by
the proc_dointvec handler.

So, the variable should be defined as unsigned int that is large
enough to use it.

+++ include/linux/sched/sysctl.h
@@ -85,7 +85,7 @@ extern int sched_group_identity_enable_handler(struct ctl_table *table, int writ
 		loff_t *ppos);
 #ifdef CONFIG_SCHED_SMT
 extern int sysctl_sched_expel_idle_balance_delay;
-extern unsigned long sysctl_sched_expel_update_interval;
+extern unsigned int sysctl_sched_expel_update_interval;
 #endif
 #endif
